requirements.txt
asgiref==3.7.2
distlib==0.3.8
Django==5.0.1
django-admin==2.0.2
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.14.0
excel-base==1.0.4
filelock==3.13.1
isoweek==1.3.3
pillow==10.2.0
platformdirs==4.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
tzlocal==5.2
virtualenv==20.25.0
xlwt==1.3.0
